Migrate standard decorators to standard provider
