+++ airflow-core/src/airflow/decorators/__init__.py
@@ -16,20 +16,11 @@
 # under the License.
 from __future__ import annotations
 
-from typing import Any, Callable
+from typing import Callable
 
 from airflow.decorators.base import TaskDecorator
-from airflow.decorators.bash import bash_task
-from airflow.decorators.branch_external_python import branch_external_python_task
-from airflow.decorators.branch_python import branch_task
-from airflow.decorators.branch_virtualenv import branch_virtualenv_task
 from airflow.decorators.condition import run_if, skip_if
-from airflow.decorators.external_python import external_python_task
-from airflow.decorators.python import python_task
-from airflow.decorators.python_virtualenv import virtualenv_task
-from airflow.decorators.sensor import sensor_task
 from airflow.decorators.setup_teardown import setup_task, teardown_task
-from airflow.decorators.short_circuit import short_circuit_task
 from airflow.decorators.task_group import task_group
 from airflow.models.dag import dag
 from airflow.providers_manager import ProvidersManager
@@ -41,15 +32,6 @@ __all__ = [
     "dag",
     "task",
     "task_group",
-    "python_task",
-    "virtualenv_task",
-    "external_python_task",
-    "branch_task",
-    "branch_virtualenv_task",
-    "branch_external_python_task",
-    "short_circuit_task",
-    "sensor_task",
-    "bash_task",
     "setup",
     "teardown",
 ]
@@ -58,19 +40,8 @@ __all__ = [
 class TaskDecoratorCollection:
     """Implementation to provide the ``@task`` syntax."""
 
-    python = staticmethod(python_task)
-    virtualenv = staticmethod(virtualenv_task)
-    external_python = staticmethod(external_python_task)
-    branch = staticmethod(branch_task)
-    branch_virtualenv = staticmethod(branch_virtualenv_task)
-    branch_external_python = staticmethod(branch_external_python_task)
-    short_circuit = staticmethod(short_circuit_task)
-    sensor = staticmethod(sensor_task)
-    bash = staticmethod(bash_task)
     run_if = staticmethod(run_if)
     skip_if = staticmethod(skip_if)
-
-    __call__: Any = python  # Alias '@task' to '@task.python'.
 
     def __getattr__(self, name: str) -> TaskDecorator:
         """Dynamically get provider-registered task decorators, e.g. ``@task.docker``."""
@@ -80,6 +51,10 @@ class TaskDecoratorCollection:
         if name not in decorators:
             raise AttributeError(f"task decorator {name!r} not found")
         return decorators[name]
+
+    def __call__(self, *args, **kwargs):
+        """Alias '@task' to python dynamically."""
+        return self.__getattr__("python")(*args, **kwargs)
 
 
 task = TaskDecoratorCollection()
