+++ airflow-core/src/airflow/decorators/__init__.pyi
@@ -28,16 +28,7 @@ from docker.types import Mount
 from kubernetes.client import models as k8s
 
 from airflow.decorators.base import FParams, FReturn, Task, TaskDecorator, _TaskDecorator
-from airflow.decorators.bash import bash_task
-from airflow.decorators.branch_external_python import branch_external_python_task
-from airflow.decorators.branch_python import branch_task
-from airflow.decorators.branch_virtualenv import branch_virtualenv_task
 from airflow.decorators.condition import AnyConditionFunc
-from airflow.decorators.external_python import external_python_task
-from airflow.decorators.python import python_task
-from airflow.decorators.python_virtualenv import virtualenv_task
-from airflow.decorators.sensor import sensor_task
-from airflow.decorators.short_circuit import short_circuit_task
 from airflow.decorators.task_group import task_group
 from airflow.models.dag import dag
 from airflow.providers.cncf.kubernetes.secret import Secret
@@ -50,15 +41,6 @@ __all__ = [
     "dag",
     "task",
     "task_group",
-    "python_task",
-    "virtualenv_task",
-    "external_python_task",
-    "branch_task",
-    "branch_virtualenv_task",
-    "branch_external_python_task",
-    "short_circuit_task",
-    "sensor_task",
-    "bash_task",
     "setup",
     "teardown",
 ]
