+++ airflow-core/src/airflow/decorators/setup_teardown.py
@@ -19,14 +19,19 @@ from __future__ import annotations
 import types
 from typing import TYPE_CHECKING, Callable
 
-from airflow.decorators import python_task
 from airflow.decorators.task_group import _TaskGroupFactory
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
+from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS
 from airflow.utils.setup_teardown import SetupTeardownContext
 
 if TYPE_CHECKING:
     from airflow import XComArg
+
+if AIRFLOW_V_3_0_PLUS:
+    from airflow.providers.standard.decorators.python import python_task
+else:
+    from airflow.decorators import python_task  # type: ignore
 
 
 def setup_task(func: Callable) -> Callable:
