--- airflow-core/tests/unit/decorators/test_branch_python.py
@@ -1,69 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from __future__ import annotations
-
-import pytest
-
-from airflow.decorators import task
-from airflow.exceptions import DownstreamTasksSkipped
-
-pytestmark = pytest.mark.db_test
-
-
-class TestBranchPythonDecoratedOperator:
-    # when run in "Parallel" test run environment, sometimes this test runs for a long time
-    # because creating virtualenv and starting new Python interpreter creates a lot of IO/contention
-    # possibilities. So we are increasing the timeout for this test to 3x of the default timeout
-    @pytest.mark.execution_timeout(180)
-    @pytest.mark.parametrize(
-        ["branch_task_name", "skipped_task_name"], [("task_1", "task_2"), ("task_2", "task_1")]
-    )
-    def test_branch_one(self, dag_maker, branch_task_name, skipped_task_name):
-        @task
-        def dummy_f():
-            pass
-
-        @task
-        def task_1():
-            pass
-
-        @task
-        def task_2():
-            pass
-
-        @task.branch(task_id="branching")
-        def branch_operator():
-            return branch_task_name
-
-        with dag_maker():
-            branchoperator = branch_operator()
-            df = dummy_f()
-            task_1 = task_1()
-            task_2 = task_2()
-
-            df.set_downstream(branchoperator)
-            branchoperator.set_downstream(task_1)
-            branchoperator.set_downstream(task_2)
-
-        dr = dag_maker.create_dagrun()
-        df.operator.run(start_date=dr.logical_date, end_date=dr.logical_date, ignore_ti_state=True)
-        with pytest.raises(DownstreamTasksSkipped) as exc_info:
-            branchoperator.operator.run(
-                start_date=dr.logical_date, end_date=dr.logical_date, ignore_ti_state=True
-            )
-        assert exc_info.value.tasks == [(skipped_task_name, -1)]
