+++ airflow-core/tests/unit/serialization/test_dag_serialization.py
@@ -2787,7 +2787,7 @@ def test_taskflow_expand_serde():
         "_is_empty": False,
         "_is_mapped": True,
         "_can_skip_downstream": False,
-        "_task_module": "airflow.decorators.python",
+        "_task_module": "airflow.providers.standard.decorators.python",
         "task_type": "_PythonDecoratedOperator",
         "_operator_name": "@task",
         "downstream_task_ids": [],
@@ -2902,7 +2902,7 @@ def test_taskflow_expand_kwargs_serde(strict):
         "_is_empty": False,
         "_is_mapped": True,
         "_can_skip_downstream": False,
-        "_task_module": "airflow.decorators.python",
+        "_task_module": "airflow.providers.standard.decorators.python",
         "task_type": "_PythonDecoratedOperator",
         "_operator_name": "@task",
         "python_callable_name": qualname(x),
