+++ providers/standard/src/airflow/providers/standard/decorators/branch_python.py
@@ -19,25 +19,25 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Callable
 
 from airflow.decorators.base import task_decorator_factory
-from airflow.decorators.python import _PythonDecoratedOperator
-from airflow.providers.standard.operators.python import BranchExternalPythonOperator
+from airflow.providers.standard.decorators.python import _PythonDecoratedOperator
+from airflow.providers.standard.operators.python import BranchPythonOperator
 
 if TYPE_CHECKING:
     from airflow.decorators.base import TaskDecorator
 
 
-class _BranchExternalPythonDecoratedOperator(_PythonDecoratedOperator, BranchExternalPythonOperator):
+class _BranchPythonDecoratedOperator(_PythonDecoratedOperator, BranchPythonOperator):
     """Wraps a Python callable and captures args/kwargs when called for execution."""
 
-    template_fields = BranchExternalPythonOperator.template_fields
-    custom_operator_name: str = "@task.branch_external_python"
+    template_fields = BranchPythonOperator.template_fields
+    custom_operator_name: str = "@task.branch"
 
 
-def branch_external_python_task(
+def branch_task(
     python_callable: Callable | None = None, multiple_outputs: bool | None = None, **kwargs
 ) -> TaskDecorator:
     """
-    Wrap a python function into a BranchExternalPythonOperator.
+    Wrap a python function into a BranchPythonOperator.
 
     For more information on how to use this operator, take a look at the guide:
     :ref:`concepts:branching`
@@ -52,6 +52,6 @@ def branch_external_python_task(
     return task_decorator_factory(
         python_callable=python_callable,
         multiple_outputs=multiple_outputs,
-        decorated_operator_class=_BranchExternalPythonDecoratedOperator,
+        decorated_operator_class=_BranchPythonDecoratedOperator,
         **kwargs,
     )
