+++ providers/standard/tests/unit/standard/decorators/test_external_python.py
@@ -29,6 +29,8 @@ import pytest
 from airflow.decorators import setup, task, teardown
 from airflow.utils import timezone
 
+from tests_common.test_utils.version_compat import AIRFLOW_V_2_10_PLUS
+
 pytestmark = pytest.mark.db_test
 
 
@@ -66,6 +68,7 @@ def venv_python_with_cloudpickle_and_dill(tmp_path_factory):
 
 
 class TestExternalPythonDecorator:
+    @pytest.mark.skipif(not AIRFLOW_V_2_10_PLUS, reason="serializer support came in after 2.10")
     @pytest.mark.parametrize(
         "serializer",
         [
@@ -86,6 +89,7 @@ class TestExternalPythonDecorator:
 
         ret.operator.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE)
 
+    @pytest.mark.skipif(not AIRFLOW_V_2_10_PLUS, reason="serializer support came in after 2.10")
     @pytest.mark.parametrize(
         "serializer",
         [
@@ -111,6 +115,7 @@ class TestExternalPythonDecorator:
 
         ret.operator.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE)
 
+    @pytest.mark.skipif(not AIRFLOW_V_2_10_PLUS, reason="serializer support came in after 2.10")
     @pytest.mark.parametrize(
         "serializer",
         [
@@ -142,6 +147,7 @@ class TestExternalPythonDecorator:
         with pytest.raises(CalledProcessError):
             ret.operator.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE)
 
+    @pytest.mark.skipif(not AIRFLOW_V_2_10_PLUS, reason="serializer support came in after 2.10")
     @pytest.mark.parametrize(
         "serializer",
         [
