+++ providers/standard/tests/unit/standard/decorators/test_sensor.py
