+++ providers/standard/tests/unit/standard/decorators/test_short_circuit.py
@@ -21,9 +21,12 @@ import pytest
 from pendulum import datetime
 
 from airflow.decorators import task
-from airflow.exceptions import DownstreamTasksSkipped
+from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS
 from airflow.utils.state import State
 from airflow.utils.trigger_rule import TriggerRule
+
+if AIRFLOW_V_3_0_PLUS:
+    from airflow.exceptions import DownstreamTasksSkipped
 
 pytestmark = pytest.mark.db_test
 
@@ -31,8 +34,57 @@ pytestmark = pytest.mark.db_test
 DEFAULT_DATE = datetime(2022, 8, 17)
 
 
+@pytest.mark.skipif(AIRFLOW_V_3_0_PLUS, reason="Test doesn't run on AF3. Companion test below.")
+def test_short_circuit_decorator_af2(dag_maker):
+    with dag_maker(serialized=True):
+
+        @task
+        def empty(): ...
+
+        @task.short_circuit()
+        def short_circuit(condition):
+            return condition
+
+        short_circuit_false = short_circuit.override(task_id="short_circuit_false")(condition=False)
+        task_1 = empty.override(task_id="task_1")()
+        short_circuit_false >> task_1
+
+        short_circuit_true = short_circuit.override(task_id="short_circuit_true")(condition=True)
+        task_2 = empty.override(task_id="task_2")()
+        short_circuit_true >> task_2
+
+        short_circuit_respect_trigger_rules = short_circuit.override(
+            task_id="short_circuit_respect_trigger_rules", ignore_downstream_trigger_rules=False
+        )(condition=False)
+        task_3 = empty.override(task_id="task_3")()
+        task_4 = empty.override(task_id="task_4")()
+        task_5 = empty.override(task_id="task_5", trigger_rule=TriggerRule.ALL_DONE)()
+        short_circuit_respect_trigger_rules >> [task_3, task_4] >> task_5
+
+    dr = dag_maker.create_dagrun()
+
+    for t in dag_maker.dag.tasks:
+        t.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+
+    task_state_mapping = {
+        "short_circuit_false": State.SUCCESS,
+        "task_1": State.SKIPPED,
+        "short_circuit_true": State.SUCCESS,
+        "task_2": State.SUCCESS,
+        "short_circuit_respect_trigger_rules": State.SUCCESS,
+        "task_3": State.SKIPPED,
+        "task_4": State.SKIPPED,
+        "task_5": State.SUCCESS,
+    }
+
+    tis = dr.get_task_instances()
+    for ti in tis:
+        assert ti.state == task_state_mapping[ti.task_id]
+
+
+@pytest.mark.skipif(not AIRFLOW_V_3_0_PLUS, reason="Test only runs on AF3")
 @pytest.mark.parametrize(["condition", "should_be_skipped"], [(True, False), (False, True)])
-def test_short_circuit_decorator(dag_maker, session, condition, should_be_skipped):
+def test_short_circuit_decorator_af3(dag_maker, session, condition, should_be_skipped):
     with dag_maker(serialized=True, session=session):
 
         @task.short_circuit()
@@ -60,6 +112,7 @@ def test_short_circuit_decorator(dag_maker, session, condition, should_be_skippe
         ti_sc.run()
 
 
+@pytest.mark.skipif(not AIRFLOW_V_3_0_PLUS, reason="Test only runs on AF3")
 @pytest.mark.parametrize(
     ["ignore_downstream_trigger_rules", "expected"], [(True, State.SKIPPED), (False, State.SUCCESS)]
 )
